Finally fixed compile errors. Fixing filters not working properly (currently they're always "passed")
