+++ SPID/include/LookupFilters.h
@@ -21,7 +21,7 @@ namespace Filter
 		std::string value;
 
 	protected:
-		StringValue(std::string value) :
+		StringValue(const std::string value) :
 			value(value) {}
 	};
 
@@ -29,7 +29,7 @@ namespace Filter
 	/// The value must be without asterisks (e.g. filter "*Vampire" should be trimmed to "Vampire")
 	struct Wildcard : StringValue
 	{
-		Wildcard(std::string value) :
+		Wildcard(const std::string value) :
 			StringValue(value) {}
 	};
 
@@ -37,7 +37,7 @@ namespace Filter
 	/// The value is stored as-is.
 	struct Match : StringValue
 	{
-		Match(std::string value) :
+		Match(const std::string value) :
 			StringValue(value) {}
 	};
 
@@ -45,15 +45,15 @@ namespace Filter
 	struct LevelRange
 	{
 		using Level = std::uint8_t;
-		inline const static Level MinLevel = 0;
-		inline const static Level MaxLevel = UINT8_MAX;
+		inline constexpr static Level MinLevel = 0;
+		inline constexpr static Level MaxLevel = UINT8_MAX;
 
 		Level min;
 		Level max;
 
-		LevelRange(Level min, Level max)
+		LevelRange(const Level min, const Level max)
 		{
-			auto _min = min == MaxLevel ? MinLevel : min;
+            const auto _min = min == MaxLevel ? MinLevel : min;
 			this->min = std::min(_min, max);
 			this->max = std::max(_min, max);
 		}
@@ -67,7 +67,7 @@ namespace Filter
 
 		Skill skill;
 
-		SkillLevelRange(Skill skill, Level min, Level max) :
+		SkillLevelRange(const Skill skill, const Level min, const Level max) :
 			LevelRange(min, max)
 		{
 			this->skill = skill;
@@ -77,7 +77,7 @@ namespace Filter
 	/// Similar to SkillLevelRange, but operates with Skill Weights of NPC's Class instead.
 	struct SkillWeightRange : SkillLevelRange
 	{
-		SkillWeightRange(Skill skill, Level min, Level max) :
+		SkillWeightRange(const Skill skill, const Level min, const Level max) :
 			SkillLevelRange(skill, min, max) {}
 	};
 
@@ -113,7 +113,7 @@ namespace Filter
 	};
 }
 
-// ---------------- Expressions ----------------
+// ------- Specialized filter evaluator --------
 namespace Filter
 {
 	/// Evaluator that determines whether given NPC matches specified filter.
@@ -127,36 +127,302 @@ namespace Filter
 		}
 	};
 
-	/// An abstract filter component that can be evaluated for specified NPC.
+	// ---------------- Strings ----------------
+
+	template <>
+	struct filter_eval<Match>
+	{
+		static Result evaluate(const Match filter, const NPCData& a_npcData)
+		{
+			if (a_npcData.GetKeywords().contains(filter.value) ||
+				string::iequals(a_npcData.GetName(), filter.value) ||
+				string::iequals(a_npcData.GetOriginalEDID(), filter.value) ||
+				string::iequals(a_npcData.GetTemplateEDID(), filter.value)) {
+				return Result::kPass;
+			}
+
+			return Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<Wildcard>
+	{
+		static Result evaluate(const Wildcard filter, const NPCData& a_npcData)
+		{
+			// Utilize regex here. (will also support one-sided wildcards (e.g. *Name and Name*)
+			// std::regex regex(filter.value, std::regex_constants::icase);
+			// std::regex_match(keyword, regex);
+			if (std::ranges::any_of(a_npcData.GetKeywords(), [&](auto keyword) { return string::icontains(keyword, filter.value); }) ||
+				string::icontains(a_npcData.GetName(), filter.value) ||
+				string::icontains(a_npcData.GetOriginalEDID(), filter.value) ||
+				string::icontains(a_npcData.GetTemplateEDID(), filter.value)) {
+				return Result::kPass;
+			}
+			return Result::kFail;
+		}
+	};
+
+	// ---------------- Forms ----------------
+
+	template <>
+	struct filter_eval<FormOrMod>
+	{
+		static Result evaluate(const FormOrMod filter, const NPCData& a_npcData)
+		{
+			if (std::holds_alternative<RE::TESForm*>(filter)) {
+				if (const auto form = std::get<RE::TESForm*>(filter); form && has_form(form, a_npcData)) {
+					return Result::kPass;
+				}
+			} else if (std::holds_alternative<const RE::TESFile*>(filter)) {
+				if (const auto file = std::get<const RE::TESFile*>(filter); file && (file->IsFormInMod(a_npcData.GetOriginalFormID()) || file->IsFormInMod(a_npcData.GetTemplateFormID()))) {
+					return Result::kPass;
+				}
+			}
+			return Result::kFail;
+		}
+
+	private:
+		/// Determines whether given form is associated with an NPC.
+		static bool has_form(RE::TESForm* a_form, const NPCData& a_npcData)
+		{
+			const auto npc = a_npcData.GetNPC();
+			switch (a_form->GetFormType()) {
+			case RE::FormType::CombatStyle:
+				return npc->GetCombatStyle() == a_form;
+			case RE::FormType::Class:
+				return npc->npcClass == a_form;
+			case RE::FormType::Faction:
+				{
+					const auto faction = a_form->As<RE::TESFaction>();
+					return npc->IsInFaction(faction);
+				}
+			case RE::FormType::Race:
+				return a_npcData.GetRace() == a_form;
+			case RE::FormType::Outfit:
+				return npc->defaultOutfit == a_form;
+			case RE::FormType::NPC:
+				{
+					const auto filterFormID = a_form->GetFormID();
+					return npc == a_form || a_npcData.GetOriginalFormID() == filterFormID || a_npcData.GetTemplateFormID() == filterFormID;
+				}
+			case RE::FormType::VoiceType:
+				return npc->voiceType == a_form;
+			case RE::FormType::Spell:
+				{
+					const auto spell = a_form->As<RE::SpellItem>();
+					return npc->GetSpellList()->GetIndex(spell).has_value();
+				}
+			case RE::FormType::Armor:
+				return npc->skin == a_form;
+			case RE::FormType::FormList:
+				{
+					bool result = false;
+
+					const auto list = a_form->As<RE::BGSListForm>();
+					list->ForEachForm([&](RE::TESForm& a_formInList) {
+						if (result = has_form(&a_formInList, a_npcData); result) {
+							return RE::BSContainer::ForEachResult::kStop;
+						}
+						return RE::BSContainer::ForEachResult::kContinue;
+					});
+
+					return result;
+				}
+			default:
+				return false;
+			}
+		}
+	};
+
+	// ---------------- Levels ----------------
+
+	template <>
+	struct filter_eval<LevelRange>
+	{
+		static Result evaluate(const LevelRange filter, const NPCData& a_npcData)
+		{
+			const auto actorLevel = a_npcData.GetLevel();
+			if (actorLevel >= filter.min && actorLevel <= filter.max) {
+				return Result::kPass;
+			}
+			return Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<SkillLevelRange>
+	{
+		static Result evaluate(const SkillLevelRange filter, const NPCData& a_npcData)
+		{
+			const auto npc = a_npcData.GetNPC();
+			const auto skillLevel = npc->playerSkills.values[filter.skill];
+			if (skillLevel >= filter.min && skillLevel <= filter.max) {
+				return Result::kPass;
+			}
+			return Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<SkillWeightRange>
+	{
+		static Result evaluate(const SkillLevelRange filter, const NPCData& a_npcData)
+		{
+			if (auto skillWeight = weight(filter, a_npcData); skillWeight >= filter.min && skillWeight <= filter.max) {
+				return Result::kPass;
+			}
+			return Result::kFail;
+		}
+
+	private:
+		static std::optional<std::uint8_t> weight(const SkillLevelRange filter, const NPCData& a_npcData)
+		{
+			if (const auto npcClass = a_npcData.GetNPC()->npcClass) {
+				const auto& skillWeights = npcClass->data.skillWeights;
+				using Skills = SkillLevelRange::Skills;
+				switch (filter.skill) {
+				case Skills::kOneHanded:
+					return skillWeights.oneHanded;
+				case Skills::kTwoHanded:
+					return skillWeights.twoHanded;
+				case Skills::kMarksman:
+					return skillWeights.archery;
+				case Skills::kBlock:
+					return skillWeights.block;
+				case Skills::kSmithing:
+					return skillWeights.smithing;
+				case Skills::kHeavyArmor:
+					return skillWeights.heavyArmor;
+				case Skills::kLightArmor:
+					return skillWeights.lightArmor;
+				case Skills::kPickpocket:
+					return skillWeights.pickpocket;
+				case Skills::kLockpicking:
+					return skillWeights.lockpicking;
+				case Skills::kSneak:
+					return skillWeights.sneak;
+				case Skills::kAlchemy:
+					return skillWeights.alchemy;
+				case Skills::kSpecchcraft:
+					return skillWeights.speech;
+				case Skills::kAlteration:
+					return skillWeights.alteration;
+				case Skills::kConjuration:
+					return skillWeights.conjuration;
+				case Skills::kDestruction:
+					return skillWeights.destruction;
+				case Skills::kIllusion:
+					return skillWeights.illusion;
+				case Skills::kRestoration:
+					return skillWeights.restoration;
+				case Skills::kEnchanting:
+					return skillWeights.enchanting;
+				default:
+					return std::nullopt;
+				}
+			}
+			return std::nullopt;
+		}
+	};
+
+	// ---------------- Traits ----------------
+
+	template <>
+	struct filter_eval<SexTrait>
+	{
+		static Result evaluate([[maybe_unused]] const SexTrait filter, const NPCData& a_npcData)
+		{
+			return a_npcData.GetSex() == filter.sex ? Result::kPass : Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<UniqueTrait>
+	{
+		static Result evaluate([[maybe_unused]] const UniqueTrait filter, const NPCData& a_npcData)
+		{
+			return a_npcData.IsUnique() ? Result::kPass : Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<SummonableTrait>
+	{
+		static Result evaluate([[maybe_unused]] const SummonableTrait filter, const NPCData& a_npcData)
+		{
+			return a_npcData.IsSummonable() ? Result::kPass : Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<ChildTrait>
+	{
+		static Result evaluate([[maybe_unused]] const ChildTrait filter, const NPCData& a_npcData)
+		{
+			return a_npcData.IsChild() ? Result::kPass : Result::kFail;
+		}
+	};
+
+	template <>
+	struct filter_eval<Chance>
+	{
+		inline static RNG staticRNG;
+
+		static Result evaluate(const Chance filter, [[maybe_unused]] const NPCData& a_npcData)
+		{
+			if (filter.maximum >= 100) {
+				return Result::kPass;
+			}
+
+			const auto rng = staticRNG.Generate<Chance::chance>(0, 100);
+			return rng > filter.maximum ? Result::kPass : Result::kFailRNG;
+		}
+	};
+}
+
+// ---------------- Expressions ----------------
+namespace Filter
+{
+    /// An abstract filter component that can be evaluated for specified NPC.
 	struct Evaluatable
 	{
-		/// Evaluates whether specified NPC matches conditions defined by this Evaluatable.
-		virtual Result evaluate([[maybe_unused]] const NPCData& a_npcData) const
+	    virtual ~Evaluatable() = default;
+
+        /// Evaluates whether specified NPC matches conditions defined by this Evaluatable.
+        [[nodiscard]] virtual Result evaluate([[maybe_unused]] const NPCData& a_npcData) const
 		{
 			return Result::kFail;
 		}
 	};
 
-	/// NegatedFilter is an entry that is always inverts the result of evaluation.
+    template <typename T>
+	struct inherits_evaluatable
+	{
+		static constexpr bool value = std::is_base_of_v<Evaluatable, T>;
+	};
+
+	/// FilterEntry evaluates whether or not given NPC matches entry's value.
+	///	It uses filter_eval specialized to FilterType.
 	template <class FilterType>
 	struct FilterEntry : Evaluatable
 	{
-		FilterType value;
+		const FilterType value;
 
-		FilterEntry(FilterType value) :
+		FilterEntry(const FilterType& value) :
 			value(value) {}
 
-		// Probably should be deleted?
-		FilterEntry<FilterType>& operator=(const FilterEntry<FilterType>&) = default;
-		FilterEntry<FilterType>& operator=(FilterEntry<FilterType>&&) = default;
+		FilterEntry(const FilterEntry& other) :
+			value(other.value) {}
 
-		virtual Result evaluate(const NPCData& a_npcData) const override
+        [[nodiscard]] Result evaluate(const NPCData& a_npcData) const override
 		{
 			return filter_eval<FilterType>::evaluate(value, a_npcData);
 		}
-	};
+    };
 
-	/// NegatedFilter is a filter that always inverts the result of evaluation.
+
+    /// NegatedFilter is a filter that always inverts the result of evaluation.
 	/// It corresponds to `-` prefix in a config file (e.g. "-ActorTypeNPC")
 	template <class FilterType>
 	struct NegatedFilter : FilterEntry<FilterType>
@@ -164,9 +430,9 @@ namespace Filter
 		NegatedFilter(FilterType value) :
 			FilterEntry<FilterType>(value) {}
 
-		virtual Result evaluate(const NPCData& a_npcData) const override
+        [[nodiscard]] Result evaluate(const NPCData& a_npcData) const override
 		{
-			switch (filter_eval<FilterType>::evaluate(FilterEntry<FilterType>::value, a_npcData)) {
+			switch (FilterEntry<FilterType>::evaluate(a_npcData)) {
 			case Result::kFail:
 			case Result::kFailRNG:
 				return Result::kPass;
@@ -176,21 +442,39 @@ namespace Filter
 			}
 		}
 	};
-
+	
 	/// An expression is a combination of filters and/or nested expressions.
 	///
 	/// To determine how entries in the expression are combined use either AndExpression or OrExpression.
 	struct Expression : Evaluatable
 	{
-		std::vector<Evaluatable> entries{};
+		std::string name;
+
+		explicit Expression(const std::string& name) :
+			name(name) {}
+
+		template <typename T, typename = std::enable_if_t<inherits_evaluatable<T>::value>>
+		void emplace_back(T* ptr)
+		{
+			static_assert(std::is_constructible_v<std::shared_ptr<Evaluatable>, T*>, "T* must be constructible into shared_ptr<Evaluatable>.");
+			entries.emplace_back(std::make_shared<T>(*ptr));
+		}
+
+		template <typename T, typename = std::enable_if_t<inherits_evaluatable<T>::value>>
+		void emplace_back(T&& obj)
+		{
+			// Not sure why, but this assertion fails, even though the method works. (this is ChatGPT stuff, so I'm not sure what exactly is wrong :))
+			//static_assert(std::is_constructible_v<std::shared_ptr<Evaluatable>, T&&>, "T must be constructible into shared_ptr<Evaluatable>.");
+			entries.emplace_back(std::make_shared<std::decay_t<T>>(std::forward<T>(obj)));
+		}
 
 		template <class FilterType>
 		void for_each_filter(std::function<void(const FilterType&)> a_callback) const
 		{
-			for (auto& eval : entries) {
-				if (auto filter = static_cast<const Expression*>(&eval)) {
-					filter->for_each_filter<FilterType>(a_callback);
-				} else if (auto entry = static_cast<const FilterEntry<FilterType>*>(&eval)) {
+			for (const auto& eval : entries) {
+				if (auto expression = dynamic_cast<Expression*>(eval.get())) {
+					expression->for_each_filter<FilterType>(a_callback);
+				} else if (auto entry = dynamic_cast<FilterEntry<FilterType>*>(eval.get())) {
 					a_callback(entry->value);
 				}
 			}
@@ -199,24 +483,31 @@ namespace Filter
 		template <class FilterType>
 		bool contains(std::function<bool(const FilterType&)> comparator) const
 		{
-			for (auto& eval : entries) {
-				if (auto filter = static_cast<const Expression*>(&eval); filter->contains<FilterType>(comparator)) {
-					return true;
-				} else if (auto entry = static_cast<const FilterEntry<FilterType>*>(&eval); comparator(entry->value)) {
-					return true;
+			return std::ranges::any_of(entries, [&](const auto& eval) {
+				if (auto filter = dynamic_cast<Expression*>(eval.get())) {
+					return filter->contains<FilterType>(comparator);
 				}
-			}
-			return false;
+				if (auto entry = dynamic_cast<FilterEntry<FilterType>*>(eval.get())) {
+					return comparator(entry->value);
+				}
+				return false;
+			});
 		}
+
+	protected:
+		std::vector<std::shared_ptr<Evaluatable>> entries{};
 	};
 
 	/// Entries are combined using AND logic.
 	struct AndExpression : Expression
 	{
-		virtual Result evaluate(const NPCData& a_npcData) const override
+		 explicit AndExpression(const std::string& name) :
+			Expression(name) {}
+
+        [[nodiscard]] Result evaluate(const NPCData& a_npcData) const override
 		{
 			for (auto& entry : entries) {
-				auto res = entry.evaluate(a_npcData);
+				const auto res = entry.get()->evaluate(a_npcData);
 				if (res != Result::kPass) {
 					return res;
 				}
@@ -228,20 +519,24 @@ namespace Filter
 	/// Entries are combined using OR logic.
 	struct OrExpression : Expression
 	{
-		virtual Result evaluate(const NPCData& a_npcData) const override
+		explicit OrExpression(const std::string& name) :
+			Expression(name) {}
+
+        [[nodiscard]] Result evaluate(const NPCData& a_npcData) const override
 		{
 			Result failure = Result::kFail;
 
 			for (auto& entry : entries) {
-				auto res = entry.evaluate(a_npcData);
+                const auto res = entry.get()->evaluate(a_npcData);
 				if (res == Result::kPass) {
 					return Result::kPass;
-				} else if (res > failure) {
-					// we rely on Result cases being ordered in order of priotities.
-					// Hence if there was a kFailRNG it will always be the result of this evaluation.
-					failure = res;
 				}
-			}
+                if (res > failure) {
+                    // we rely on Result cases being ordered in order of priorities.
+                    // Hence if there was a kFailRNG it will always be the result of this evaluation.
+                    failure = res;
+                }
+            }
 			return entries.empty() ? Result::kPass : failure;
 		}
 	};
@@ -266,20 +561,3 @@ namespace Filter
 }
 
 using FilterData = Filter::Data;
-
-/*
-Input: key1,-key2,-key3+key4,-*key5+key6+-key7
-Step 1: split ","
- key1
- -key2
- -key3+key4
- -*key5+key6+-key7
-
- Step 2: split "+"
- -key3
- key4
-
- -*key5
- key6
- -key7
-*/
